Move files into flux_based/ and cell_based/ folders
