+++ include/samurai/schemes/fv/flux_based/flux_based_scheme.hpp
@@ -1,7 +1,7 @@
 #pragma once
-#include "../../interface.hpp"
-#include "../explicit_scheme.hpp"
-#include "FV_scheme.hpp"
+#include "../../../interface.hpp"
+#include "../../explicit_scheme.hpp"
+#include "../FV_scheme.hpp"
 #include "flux_definition.hpp"
 
 namespace samurai
