+++ include/samurai/schemes/fv/flux_based/flux_definition.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "utils.hpp"
+#include "../utils.hpp"
 #include <functional>
 
 namespace samurai
