+++ include/samurai/schemes/fv/operators/diffusion_FV.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include "../flux_based_scheme__lin_het.hpp"
-#include "../flux_based_scheme__lin_hom.hpp"
+#include "../flux_based/flux_based_scheme__lin_het.hpp"
+#include "../flux_based/flux_based_scheme__lin_hom.hpp"
 #include "divergence_FV.hpp"
 
 namespace samurai
