+++ include/samurai/schemes/fv/operators/divergence_FV.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../flux_based_scheme__lin_hom.hpp"
+#include "../flux_based/flux_based_scheme__lin_hom.hpp"
 
 namespace samurai
 {
